[AmILogged] Change logic from app to header component
