+++ src/app/app.component.ts
@@ -1,6 +1,4 @@
 import { Component, OnInit } from '@angular/core';
-import moment from 'moment-timezone';
-import { environment } from '../environments/environment';
 
 @Component({
   selector: 'app-root',
@@ -8,20 +6,7 @@ import { environment } from '../environments/environment';
   styleUrls: ['../style.scss']
 })
 export class AppComponent implements OnInit {
-  localStorageUserKey = 'shareBookUser';
 
   ngOnInit() {
-    this.checkTokenValidity();
-  }
-
-  checkTokenValidity() {
-    const user = JSON.parse(localStorage.getItem(this.localStorageUserKey));
-    if (user) {
-      const expiration = moment(user.expiration);
-      const now = moment();
-      if (now.isAfter(expiration)) {
-        localStorage.setItem(this.localStorageUserKey, null);
-      }
-    }
   }
 }
