+++ src/app/components/header/header.component.ts
@@ -1,5 +1,7 @@
 import { Component, OnInit, ViewChild, ElementRef } from '@angular/core';
 import { Subscription } from 'rxjs';
+import moment from 'moment-timezone';
+import { environment } from '../../../environments/environment';
 
 import { User } from '../../core/models/user';
 import { UserService } from '../../core/services/user/user.service';
@@ -18,6 +20,7 @@ export class HeaderComponent implements OnInit {
 
   userLogged = false;
   shareBookUser = new User();
+  localStorageUserKey = 'shareBookUser';
 
   constructor(private _scUser: UserService) {
 
@@ -34,12 +37,24 @@ export class HeaderComponent implements OnInit {
       this.shareBookUser = shareBookUser;
       this.userLogged = !!this.shareBookUser;
     });
+    this.checkTokenValidity();
   }
 
   // metodo que desativa o menu ao clicar em um link
   showHideMenu() {
     if (this.menu.nativeElement.classList.contains('show')) {
       this.menu.nativeElement.classList.toggle('show');
+    }
+  }
+
+  checkTokenValidity() {
+    const user = JSON.parse(localStorage.getItem(this.localStorageUserKey));
+    if (user) {
+      const expiration = moment(user.expiration);
+      const now = moment();
+      if (now.isAfter(expiration)) {
+        localStorage.setItem(this.localStorageUserKey, null);
+      }
     }
   }
 }
