+++ src/environments/environment.prod.ts
@@ -1,4 +1,5 @@
 export const environment = {
   production: true,
-  apiEndpoint: `http://www.sharebook.com.br/api`
+  apiEndpoint: `http://www.sharebook.com.br/api`,
+  backendTimeZone: `America/Los_Angeles`
 };
