+++ src/environments/environment.stg.ts
@@ -1,4 +1,5 @@
 export const environment = {
   production: false,
-  apiEndpoint: `http://stg.sharebook.com.br/api`
+  apiEndpoint: `http://stg.sharebook.com.br/api`,
+  backendTimeZone: `America/Los_Angeles`
 };
