+++ src/environments/environment.ts
@@ -4,7 +4,8 @@
 
 export const environment = {
   production: false,
-  apiEndpoint: `http://dev.sharebook.com.br/api`
+  apiEndpoint: `http://dev.sharebook.com.br/api`,
+  backendTimeZone: `America/Los_Angeles`
 };
 
 /*
